WiP: Try decrypt AES-GCM
The code flow is implemented, now with tag and nonce derivation from RFC 8446, however, without success. It seems that nonce derivation is correct, but the sequence number was not extracted from pyshark and this could be an error. Or the tag is not correct. Should debug to find and/or find test cases