WiP: SequenceNumber decrypt attempt
It adds the try-and-error attempt for sequence numbers, because it was not possible to obtain sequence numbers from the capture file. So it graps app data and it tries all sequence numbers (until len(packets)). There is still fixes and now two examples in Go will be used as basis